供应链V1.0 (数据权限优化)

+++ supply-pojo/src/main/java/com/yitai/admin/dto/user/UserDTO.java
@@ -35,7 +35,7 @@ public class UserDTO extends BaseBody {
     @ExcelImport(value = "头像")
     private String avatar;
     @ExcelImport(value = "关联部门")
-    private List<Integer> deptIds;
+    private List<Long> deptIds;
     @ExcelImport(value = "关联角色")
-    private List<Integer> roleIds;
+    private List<Long> roleIds;
 }
